--- mincheol/hwcamp-coding-test-study/src/main/java/Week03/LV1/dataanalisting/DataAnalisting.java
@@ -1,12 +1,0 @@
-package Week03.LV1.dataanalisting;
-
-public class DataAnalisting {
-
-    public static void main(String[] args) {
-
-        System.out.println(Solution.solution(new int[][]{{1, 20300104, 100, 80},
-                                                         {2, 20300804, 847, 37},
-                                                         {3, 20300401, 10, 8}},
-                                                    "date", 20300501,"remain"));
-    }
-}
