--- mincheol/hwcamp-coding-test-study/src/main/java/Week03/LV1/dataanalisting/Solution.java
@@ -1,17 +1,0 @@
-package Week03.LV1.dataanalisting;
-
-/*
- * 이름: 변민철
- * 날짜: 25. 08. 19.
- * 문제: https://school.programmers.co.kr/learn/courses/30/lessons/250121
- * */
-
-public class Solution {
-    public static int[][] solution(int[][] data, String ext, int val_ext, String sort_by) {
-        int[][] answer = {};
-
-
-        return answer;
-    }
-}
-
